+++ app/Http/Controllers/DeviceController.php
@@ -88,7 +88,11 @@ class DeviceController extends Controller
     }
 
     public function delete(Request $request,$id){
-        $user =$request->user()->id;
+        $auth = $request->user();
+        $user = $auth->id;
+        if( $auth->type != "admin"){
+            return response(['message'=>'權限不足'],403);
+        }
         $data=Device::find($id);
         $this->log->deleteLog('dev',$user,$data);
         $data->delete();
